\name{covLedoit}

\alias{covLedoit}

\title{
A wrapper for SKCOv
}

\encoding{latin1}

\description{
A wrapper for SKCOv.
}

\usage{covLedoit(x,spec = NULL)
}
\arguments{
  \item{x}{Time series object}
  \item{spec}{NULL}

  }


\details{
A wrapper for SKCOv.
}

\value{
A wrapper for SKCOv.
}


\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}

