\name{covStudent}

\alias{covStudent}

\title{
A covariance estimator based on Student distritbution from package MASS
}

\encoding{latin1}

\description{
A covariance estimator based on Student distritbution from package MASS.
}

\usage{covStudent(x,spec = NULL)
}
\arguments{
  \item{x}{Time series object}
  \item{spec}{NULL}

  }


\details{
Call a covariance estimator based on Student distritbution from package MASS.
}

\value{
A covariance estimator based on Student distritbution.
}

\references{
Kent J. T., D. E. Tyler and Y. Vardi (1994). A curious likelihood identity for the multivariate t distribution.  Communications in Statistics—Simulation and Computation 23, 441–453.
}


\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}
