\name{data-sets}


\alias{dataSets}

\alias{assetReturns}

\alias{macrodata}

\title{Data Sets}

\encoding{latin1}

\description{

assetReturns contains DJ component stocks returns data. macrodata contains US unemployment(unrate) and year-to-year changes in three regional business cycle indices (OECD, NAFTA, and G7).

}


\usage{
data(assetReturns)
data(macrodata)
}


\value{

  assetReturns is a time series object of package \code{xts}.
	macrodata is macroeconomic time series data.
}


\keyword{datasets}
