\name{DJ30}
\alias{DJ30}
\docType{data}
\title{
Close Price Data of Dow Jone 30 Component
}

\encoding{latin1}

\description{
Daily close price data of DJ30, 2007/4/24~2017/4/21
}
\usage{data("DJ30")}
\format{
  A xts object with 2518 observations of the close price of DJ 30 component.

}


\details{
Daily stock price data of DJ30. Transform this dataset into DJ30.RData and use it as an example dataset for Assets Selection and Portfolio Backtesting. The code easy: data(DJ30);save(DJ30,file="DJ30.RData")
}
\source{
Yahoo finance.
}
