\name{GMVPStrategy}

\alias{GMVPStrategy}

\title{
A wrapper of GMVP strategy for backtesting a fPortfolio class
}

\encoding{latin1}

\description{
A wrapper of GMVP strategy for backtesting a fPortfolio class.
}

\usage{GMVPStrategy(data, spec, constraints, backtest)
}
\arguments{
  \item{data}{Time series object}
  \item{spec}{Specification for Porfolio optimization}
  \item{constraints}{Constraint for fPorfolio}
  \item{backtest}{Specification for backtesting}


  }


\details{
A wrapper of GMVP strategy for backtesting a fPortfolio class.
}

\value{
A wrapper of GMVP strategy for backtesting a fPortfolio class.
}


\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}


