\name{ShrinkCC}

\alias{ShrinkCC}

\title{
Shrinkage covariance estimator with constant correlation for portfolio optimization.
}

\encoding{latin1}

\description{
Shrinkage covariance estimator with constant correlation
}

\usage{ShrinkCC(x,spec = NULL)
}
\arguments{
  \item{x}{Time series object}
  \item{spec}{NULL}

  }


\details{
Generate a shrinkage covariance estimator with constant correlation for portfolio optimization.
}

\value{
Generate a shrinkage covariance estimator with constant correlation
}



\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}




