\name{SKCov}

\alias{SKCov}

\title{
A covariance estimator based on Baysian shrinkage estimator
}

\encoding{latin1}

\description{
A covariance estimator based on a Baysian shrinkage estimator.
}

\usage{SKCov(dat)
}
\arguments{
  \item{dat}{Time series object}

  }


\details{
Call a covariance estimator based on Baysian shrinkage estimator.
}

\value{
A covariance estimator based on Baysian shrinkage estimator.
}


\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}

