\name{world20}
\alias{world20}
\docType{data}
\title{
Close Price Data of twenty national market indices
}

\encoding{latin1}

\description{
Daily close price data of world20, 2007/4/24~2017/4/21
}
\usage{data("world20")}
\format{
  A xts object with 2518 observations of twenty national market indices.

}


\details{
  A xts object with 2518 observations of twenty national market indices.
}
\source{
Yahoo finance.
}
