\name{covLedoit}

\alias{covLedoit}

\title{
A wrapper for SKCOv
}

\encoding{latin1}

\description{
A wrapper for SKCOv.
}

\usage{covLedoit(x,spec = NULL)
}
\arguments{
  \item{x}{Time series object}
  \item{spec}{NULL}

  }


\details{
A wrapper for SKCOv.
}

\value{
A wrapper for SKCOv.
}

\references{
Ledoit, O. and M. Wolf, (2003) Improved estimation of the covariance matrix of returns with an application to portfolio selection.  Journal of Empirical Finance 10, 603-621.
}

\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}

