\name{data-sets}


\alias{dataSets}

\alias{assetReturns}


\title{DJ 30 Assets Data Sets}

\encoding{UTF-8}

\description{

    Example data sets.

}


\usage{
data(assetReturns)
}


%\details{
%}


\value{

    an object of class \code{"timeSeries"} or \code{"xts"}.

}


\keyword{datasets}

