\name{data-sets}


\alias{dataSets}

\alias{assetReturns}

\title{Assets Data Sets}

\encoding{UTF-8}

\description{

assetReturns contains DJ component stocks returns data.

}


\usage{
data(assetReturns)
}



\value{

    an object of class \code{"timeSeries"} or \code{"xts"}.

}


\keyword{datasets}

