\name{data-sets}


\alias{dataSets}

\alias{assetReturns}

\alias{macrodata}

\title{Assets Data Sets}

\encoding{UTF-8}

\description{

assetReturns contains DJ component stocks returns data. macrodata contains US unemployment(unrate) and year-to-year changes in three regional business cycle indices (OED, NAFTA, and G7).

}


\usage{
data(assetReturns)
data(macrodata)
}



\value{

    an object of class \code{"xts"}.

}


\keyword{datasets}

