\name{data-sets}


\alias{dataSets}

\alias{DJ30}
\alias{IBM}
\alias{world20}


\title{Assets Data Sets}


\description{

    Example data sets.

}


\usage{
DJ30
IBM
world20
}


%\details{
%}


\value{

    an object of class \code{"timeSeries"} or \code{"xts"}.

}


\keyword{datasets}

