\name{data-sets}


\alias{dataSets}

\alias{assetReturns}

\alias{macrodata}

\title{Assets Data Sets}

\encoding{UTF-8}

\description{

assetReturns contains DJ component stocks returns data. macrodata contains US unemployment(unrate) and year-to-year changes in three regional business cycle indices (OED, NAFTA, and G7).

}


\usage{
data(assetReturns)
data(macrodata)

}



\value{

    assetReturns is a time series object of package \code{"xts"}; the others are time series objects of package \code{"timeSeries"}.

}


\keyword{datasets}

