\name{GoldSach}

\alias{GoldSach}

\title{
Goldman-Sach covariance estimator for portfolio optimization.
}

\encoding{latin1}

\description{
Goldman-Sach covariance estimator with constant correlation
}

\usage{GoldSach(x,spec = NULL)
}
\arguments{
  \item{x}{Time series object}
  \item{spec}{NULL}

  }


\details{
Generate a Goldman-Sach  covariance estimator portfolio optimization.
}

\value{
Generate a Goldman-Sach covariance covariance estimator.
}


\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}
