\name{ShrinkCC}

\alias{ShrinkCC}

\title{
Ledoit-Wolf Shrinkage covariance estimator with constant correlation for portfolio optimization
}

\encoding{latin1}


\description{
Ledoit-Wolf Shrinkage covariance estimator with constant correlation
}

\usage{ShrinkCC(x,spec = NULL)
}

\arguments{
 \item{x}{Time series object}
 \item{spec}{NULL}

  }


\details{
Generate a Ledoit-Wolf shrinkage covariance estimator with constant correlation for portfolio optimization.
}

\value{
Generate a Ledoit-Wolf shrinkage covariance estimator with constant correlation
}

\references{
Ledoit, O. and M. Wolf (2004) Honey, I shrunk the sample covariance matrix. Journal of Portfolio Management 30, 110-119.
}

\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}




